worked a ton of lines that rely on undefined behaviour out of the code. New issues are popping up as a result, but that gives me a lot of good leads to chase down

+++ src/gb.hpp
@@ -23,8 +23,8 @@ class gb {
 		std::shared_ptr<ppu> _ppu;
 
 	protected:
-		unsigned long	_cycles;
-		unsigned		cyc;
+		unsigned long long	_cycles;
+		unsigned			cyc;
 };
 
 #endif
